Fixed warnings
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@836 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/xap/win/xap_Win32Frame.cpp
@@ -660,7 +660,7 @@ LRESULT CALLBACK AP_Win32Frame::_FrameWndProc(HWND hwnd, UINT iMsg, WPARAM wPara
 		int nWidth = LOWORD(lParam);
 		int nHeight = HIWORD(lParam);
 
-		if (nWidth != f->m_iSizeWidth)
+		if (nWidth != (int) f->m_iSizeWidth)
 		{
 			MoveWindow(f->m_hwndRebar, 0, 0, nWidth, f->m_iBarHeight, TRUE); 
 		}
